UHF-10566: Move block definition to paatokset_allu module
