+++ public/modules/custom/paatokset_allu/paatokset_allu.module
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * @file
+ * Primary module hooks for Paatokset allu module.
+ */
+
+declare(strict_types=1);
+
+/**
+ * Implements hook_theme().
+ */
+function paatokset_allu_theme() : array {
+  return [
+    'allu_decisions_search_block' => [
+      'variables' => [],
+      'template' => 'allu-decisions-search-block',
+    ],
+  ];
+}
