+++ public/modules/custom/paatokset_allu/src/Plugin/Block/AlluDecisionsSearchBlock.php
@@ -1,0 +1,68 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Drupal\paatokset_allu\Plugin\Block;
+
+use Drupal\Component\Plugin\ContextAwarePluginInterface;
+use Drupal\Core\Block\Attribute\Block;
+use Drupal\Core\Block\BlockBase;
+use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
+use Drupal\Core\StringTranslation\TranslatableMarkup;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+
+/**
+ * Provides Block for allu decisions search.
+ */
+#[Block(
+  id: "allu_decisions_search_block",
+  admin_label: new TranslatableMarkup("Allu decisions search block"),
+)]
+final class AlluDecisionsSearchBlock extends BlockBase implements ContainerFactoryPluginInterface, ContextAwarePluginInterface {
+
+  /**
+   * @param array $configuration
+   *   A configuration array containing information about the plugin instance.
+   * @param string $plugin_id
+   *   The plugin ID for the plugin instance.
+   * @param mixed $plugin_definition
+   *   The plugin implementation definition.
+   *  @param \Drupal\Core\Config\ConfigFactoryInterface $config
+   *   The config factory.
+   */
+  public function __construct(
+    array $configuration,
+    $plugin_id,
+    $plugin_definition,
+    private readonly ConfigFactoryInterface $config
+  ) {
+    parent::__construct($configuration, $plugin_id, $plugin_definition);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) : self {
+    return new self($configuration, $plugin_id, $plugin_definition,
+      $container->get('config.factory'),
+    );
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function build(): array {
+    return [
+      '#attached' => [
+        'drupalSettings' => [
+          'helfi_react_search' => [
+            'elastic_proxy_url' => $this->config->get('elastic_proxy.settings')->get('elastic_proxy_url'),
+          ],
+        ],
+      ],
+      '#theme' => 'allu_decisions_search_block',
+    ];
+  }
+
+}
